+++ protected/humhub/modules/space/modules/manage/views/security/permissions.php
@@ -1,5 +1,6 @@
 <?php
 
+use humhub\modules\user\widgets\PermisionGridModuleFilter;
 use humhub\modules\user\widgets\PermissionGridEditor;
 use humhub\modules\space\modules\manage\widgets\SecurityTabMenu;
 ?>
@@ -17,13 +18,20 @@ use humhub\modules\space\modules\manage\widgets\SecurityTabMenu;
         <p class="help-block"><?= Yii::t('SpaceModule.manage', 'Permissions are assigned to different user-roles. To edit a permission, select the user-role you want to edit and change the drop-down value of the given permission.'); ?></p>
     </div>
 
-    <ul id="tabs" class="nav nav-tabs tab-sub-menu">
+    <div class="clearfix">
+        <?= PermisionGridModuleFilter::widget() ?>
+    </div>
+
+    <ul id="tabs" class="nav nav-tabs tab-sub-menu permission-group-tabs">
         <?php foreach ($groups as $currentGroupId => $groupLabel) : ?>
             <li class="<?= ($groupId === $currentGroupId) ? 'active' : '' ?>">
                 <a href='<?= $space->createUrl('permissions', ['groupId' => $currentGroupId]); ?>'><?= $groupLabel ?></a>
             </li>
         <?php endforeach; ?>
     </ul>
+
+
+
     <div class="panel-body" style="padding-top: 0px;">
         <?= PermissionGridEditor::widget(['permissionManager' => $space->permissionManager, 'groupId' => $groupId]); ?>
     </div>