+++ protected/humhub/docs/RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 1.5
 ------------------------------------
 
+- Permission overview filter (https://github.com/humhub/humhub/issues/3557) (TBD)
 - Admin log filter (https://github.com/humhub/humhub/issues/3909) (TBD)
 - SameSite Cookie Support (TBD)